feat: Add implementaion for Parser.getStringOrUndefined
Add this method so that a user can parse the value from an Object's field,
given a key path, into a string OR undefined.

A StringValidationException is thrown if the field exists but is NOT a string
so that whenever there is a runtime exception, the user will know exactly what
went wrong.