feat: Add implementation for Parser.getStrings
Add this method so that the user can extract the value of an unknown
object's field and parse it into an array of strings given the full
key path of the field. If the no key path is provided then the Parser
will parse the entire object itself as an array of strings.