feat: Add implementation for Parser.getNumbersOrUndefined
Make this change so that users can extract a field from an object with
an unknown shape and parse it into an array of numbers or undefined given
a full keypath.

This is useful for parsing fields from HTTP request bodies or for raw data
returned from a SQL query.