feat: Add implementation for Parser.getBooleanOrUndefined
Add this implementation so that users can extract a value from
a field of an object with an unknown shape and parse it into a
boolean or return undefined if the field does not exist.

This is useful for parsing HTTP JSON request bodies or raw data
returned from an SQL query.